Improved Elections page and linked to it from Welcome.
